split reservations component
